refactor: move async fs api to fsPromisesPolyfill.ts
