+++ packages/utoo-web/src/utoo/index.d.ts
@@ -175,11 +175,11 @@ export interface InitOutput {
   readonly __wbindgen_export_7: WebAssembly.Table;
   readonly __externref_drop_slice: (a: number, b: number) => void;
   readonly __externref_table_dealloc: (a: number) => void;
-  readonly wasm_bindgen__convert__closures_____invoke__h2ce973260553dde0: (a: number, b: number) => void;
-  readonly closure115096_externref_shim: (a: number, b: number, c: any) => void;
   readonly closure117648_externref_shim: (a: number, b: number, c: any) => void;
   readonly closure149_externref_shim: (a: number, b: number, c: any) => void;
+  readonly wasm_bindgen__convert__closures_____invoke__h2ce973260553dde0: (a: number, b: number) => void;
   readonly wasm_bindgen__convert__closures_____invoke__heac94dfe74335608: (a: number, b: number) => void;
+  readonly closure115096_externref_shim: (a: number, b: number, c: any) => void;
   readonly closure115099_externref_shim: (a: number, b: number, c: any) => void;
   readonly closure117778_externref_shim: (a: number, b: number, c: any, d: any) => void;
   readonly __wbindgen_thread_destroy: (a?: number, b?: number, c?: number) => void;
