+++ packages/utoo-web/src/webpackLoaders/polyfills/fsPolyfill.ts
@@ -1,7 +1,7 @@
 import { Buffer } from "buffer";
 import path from "path";
 import * as sabcom from "../../sabcom";
-import { RawStats, Stats } from "../../type";
+import { Stats } from "../../type";
 
 function resolvePath(p: string): string {
   // @ts-ignore
@@ -293,89 +293,8 @@ export function access(p: string, mode: number | Function, cb?: Function) {
   });
 }
 
-export const promises = {
-  readFile: async (p: string, options?: any) => {
-    const encoding =
-      options === "utf8" ||
-      options === "utf-8" ||
-      options?.encoding === "utf8" ||
-      options?.encoding === "utf-8"
-        ? "utf8"
-        : undefined;
-    const fs = getFs();
-    const path = resolvePath(p);
-    const data = await (encoding ? fs.readToString(path) : fs.read(path));
-    return encoding ? data : Buffer.from(data);
-  },
-  writeFile: async (p: string, data: string | Uint8Array, options?: any) => {
-    const fs = getFs();
-    const path = resolvePath(p);
-    if (typeof data === "string") {
-      await fs.writeString(path, data);
-    } else {
-      await fs.write(path, data);
-    }
-  },
-  readdir: async (p: string, options?: any) => {
-    const entries = await getFs().readDir(resolvePath(p));
-    return entries.map((e: any) => {
-      const json = e.toJSON() as any;
-      if (options?.withFileTypes) {
-        return {
-          name: json.name,
-          isFile: () => json.type === "file",
-          isDirectory: () => json.type === "directory",
-          isSymbolicLink: () => false,
-        };
-      }
-      return json.name;
-    });
-  },
-  mkdir: async (p: string, options?: any) => {
-    const fs = getFs();
-    const path = resolvePath(p);
-    if (options?.recursive) {
-      await fs.createDirAll(path);
-    } else {
-      await fs.createDir(path);
-    }
-  },
-  rm: async (p: string, options?: any) => {
-    const fs = getFs();
-    const path = resolvePath(p);
-    const metadata = await fs.metadata(path);
-    const type = (metadata.toJSON() as any).type;
-    if (type === "file") {
-      await fs.removeFile(path);
-    } else {
-      await fs.removeDir(path, !!options?.recursive);
-    }
-  },
-  rmdir: async (p: string, options?: any) =>
-    getFs().removeDir(resolvePath(p), !!options?.recursive),
-  copyFile: async (src: string, dst: string) =>
-    getFs().copyFile(resolvePath(src), resolvePath(dst)),
-  stat: async (p: string) => {
-    const metadata = await getFs().metadata(resolvePath(p));
-    const json = metadata.toJSON() as any;
-    return new Stats({
-      type: json.type,
-      size: Number(json.file_size || 0),
-    });
-  },
-  lstat: async (p: string) => {
-    const metadata = await getFs().metadata(resolvePath(p));
-    const json = metadata.toJSON() as any;
-    return new Stats({
-      type: json.type,
-      size: Number(json.file_size || 0),
-    });
-  },
-  realpath: async (p: string) => p,
-  access: async (p: string, mode?: number) => {
-    await getFs().metadata(resolvePath(p));
-  },
-};
+import { promises } from "./fsPromisesPolyfill";
+export { promises };
 
 export const constants = {
   F_OK: 0,
